requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-generativeai==0.8.5
Pillow==10.1.0
pydantic==2.5.0
